requirements.txt
pyorbit-tools @ https://github.com/austin-hoover/pyorbit-tools.git
